+++ src/config.py
@@ -1,0 +1,23 @@
+from colorama import Fore, Back, Style
+
+class print_stuff:
+    def __init__(self, reset: int = 0):
+        self.reset = reset
+        self.print_stars = f"{Fore.MAGENTA}*" * 93
+        self.reset_stars = self.print_stars + Style.RESET_ALL
+
+    def printStars(self) -> None:
+        p = self.print_stars
+        if self.reset:
+            p = self.reset_stars
+        print(p)
+
+    def stringStars(self) -> str:
+        p = self.print_stars
+        if self.reset:
+            p = self.reset_stars
+        return p
+
+    @classmethod
+    def printWhitespace(self) -> None:
+        print("\n" * 8)
